pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-012"
version = "0.1.0"
description = "HW-12"
authors = ["Yulia  <nayanzina@gmail.com>"]

readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
alembic = "^1.15.1"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.39"
pydantic = "^2.10.3"
pydantic-settings = "^2.8.1"
greenlet = "^3.1.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
libgravatar = "^1.0.4"
python-dotenv = "^1.0.1"
slowapi = "^0.1.9"
redis = "^5.2.1"
redis-lru = "^0.1.2"
fastapi-mail = "^1.4.2"
cloudinary = "^1.41.0"
psycopg2-binary = "^2.9.10"
aiocache = "^0.12.3"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
aiosqlite = "^0.21.0"
pytest-cov = "^6.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
